+++ aksel.nav.no/website/pages/eksempler/datepicker/validation.tsx
@@ -3,10 +3,13 @@ import { withDsExample } from "components/website-modules/examples/withDsExample
 import { useState } from "react";
 
 const Example = () => {
-  const [day, setDay] = useState<Date | undefined>();
+  const [hasError, setHasError] = useState(false);
   const { datepickerProps, inputProps } = UNSAFE_useDatepicker({
     fromDate: new Date("Aug 23 2019"),
-    onDateChange: setDay,
+    onValidate: (val) => {
+      setHasError(!val.isValidDate);
+      console.log(val);
+    },
   });
 
   return (
@@ -15,7 +18,7 @@ const Example = () => {
         <UNSAFE_DatePicker.Input
           {...inputProps}
           label="Velg dato"
-          error={!day && "Må velge en dag"}
+          error={hasError && "Noe er feil"}
         />
       </UNSAFE_DatePicker>
     </div>
@@ -26,5 +29,5 @@ export default withDsExample(Example);
 
 export const args = {
   index: 9,
-  desc: "Vi anbefaler å bruke UNSAFE_useDatepicker-hook hvis man har et input-felt",
+  desc: "Bruk onValidate-callback for å håndtere validering.",
 };