+++ aksel.nav.no/website/pages/eksempler/dropdown/default.tsx
@@ -4,25 +4,29 @@ import { withDsExample } from "components/website-modules/examples/withDsExample
 
 const Example = () => {
   return (
-    <Dropdown>
-      <Button as={Dropdown.Toggle}>Toggle</Button>
-      <Dropdown.Menu>
-        <Dropdown.Menu.GroupedList>
-          <Dropdown.Menu.GroupedList.Heading>
-            Systemer og oppslagsverk
-          </Dropdown.Menu.GroupedList.Heading>
-          <Dropdown.Menu.GroupedList.Item>Gosys</Dropdown.Menu.GroupedList.Item>
-          <Dropdown.Menu.GroupedList.Item>
-            Infotrygd
-          </Dropdown.Menu.GroupedList.Item>
-        </Dropdown.Menu.GroupedList>
-        <Dropdown.Menu.Divider />
-        <Dropdown.Menu.List>
-          <Dropdown.Menu.List.Item>Gosys</Dropdown.Menu.List.Item>
-          <Dropdown.Menu.List.Item>Infotrygd</Dropdown.Menu.List.Item>
-        </Dropdown.Menu.List>
-      </Dropdown.Menu>
-    </Dropdown>
+    <div className="min-h-32">
+      <Dropdown>
+        <Button as={Dropdown.Toggle}>Toggle</Button>
+        <Dropdown.Menu>
+          <Dropdown.Menu.GroupedList>
+            <Dropdown.Menu.GroupedList.Heading>
+              Systemer og oppslagsverk
+            </Dropdown.Menu.GroupedList.Heading>
+            <Dropdown.Menu.GroupedList.Item>
+              Gosys
+            </Dropdown.Menu.GroupedList.Item>
+            <Dropdown.Menu.GroupedList.Item>
+              Infotrygd
+            </Dropdown.Menu.GroupedList.Item>
+          </Dropdown.Menu.GroupedList>
+          <Dropdown.Menu.Divider />
+          <Dropdown.Menu.List>
+            <Dropdown.Menu.List.Item>Gosys</Dropdown.Menu.List.Item>
+            <Dropdown.Menu.List.Item>Infotrygd</Dropdown.Menu.List.Item>
+          </Dropdown.Menu.List>
+        </Dropdown.Menu>
+      </Dropdown>
+    </div>
   );
 };