+++ aksel.nav.no/website/pages/eksempler/tag/success.tsx
@@ -2,11 +2,17 @@ import { Tag } from "@navikt/ds-react";
 import { withDsExample } from "components/website-modules/examples/withDsExample";
 
 const Example = () => {
-  return <Tag variant="success">Success</Tag>;
+  return (
+    <div className="flex gap-2">
+      <Tag variant="success">Success</Tag>
+      <Tag variant="success-filled">Success</Tag>
+    </div>
+  );
 };
 
 export default withDsExample(Example);
 
 export const args = {
-  index: 1,
+  index: 2,
+  desc: "Beskriver noe som er positivt, gjennomført, vellykket, etc.",
 };