+++ aksel.nav.no/website/pages/eksempler/tag/warning.tsx
@@ -2,11 +2,17 @@ import { Tag } from "@navikt/ds-react";
 import { withDsExample } from "components/website-modules/examples/withDsExample";
 
 const Example = () => {
-  return <Tag variant="warning">Warning</Tag>;
+  return (
+    <div className="flex gap-2">
+      <Tag variant="warning">Warning</Tag>
+      <Tag variant="warning-filled">Warning</Tag>
+    </div>
+  );
 };
 
 export default withDsExample(Example);
 
 export const args = {
-  index: 2,
+  index: 3,
+  desc: "Beskriver noe som er en advarsel, usikkert, noe som må utbedres, etc.",
 };