+++ @navikt/internal/css/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @navikt/ds-css-internal
 
+## 1.5.5
+
+## 1.5.4
+
+## 1.5.3
+
+## 1.5.2
+
+## 1.5.1
+
+## 1.5.0
+
+## 1.4.4
+
+## 1.4.3
+
+## 1.4.2
+
+### Patch Changes
+
+- :bug: Dropdown-link ikke i synk med Figma
+
+## 1.4.1
+
 ## 1.4.0
 
 ## 1.3.39